# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  icon: string;
  imageUrl: string;
  createdAt?: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  icon?: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  icon?: string;
  imageUrl?: string;
}

export interface OrderItemInput {
  /** @minimum 1 */
  productId: number;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  address: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  New: 'New',
  Processing: 'Processing',
  Shipped: 'Shipped',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  address: string;
  total: number;
  status: OrderStatus;
  createdAt: string;
  items: OrderItem[];
}

export type OrderStatusInputStatus = typeof OrderStatusInputStatus[keyof typeof OrderStatusInputStatus];


export const OrderStatusInputStatus = {
  New: 'New',
  Processing: 'Processing',
  Shipped: 'Shipped',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}

export type AdminSummaryStatusCounts = {[key: string]: number};

export interface AdminSummary {
  productCount: number;
  orderCount: number;
  revenue: number;
  statusCounts: AdminSummaryStatusCounts;
}

export interface Error {
  error: string;
}

export type ListProductsParams = {
search?: string;
};

export type ListOrdersParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};