# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MarketNest marketplace API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: orders
    description: Order and checkout operations
  - name: admin
    description: Marketplace administration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List recent orders
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update an order status
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get marketplace summary
      responses:
        "200":
          description: Catalog and order summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
components:
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, description, price, icon, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0
        icon:
          type: string
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0
        icon:
          type: string
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0
        icon:
          type: string
        imageUrl:
          type: string
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 99
    OrderItem:
      type: object
      required: [productId, productName, price, quantity]
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
    OrderInput:
      type: object
      required: [customerName, email, address, items]
      properties:
        customerName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
          minLength: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
    Order:
      type: object
      required: [id, customerName, email, phone, address, total, status, createdAt, items]
      properties:
        id:
          type: integer
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        total:
          type: number
          format: double
        status:
          type: string
          enum: [New, Processing, Shipped, Completed, Cancelled]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [New, Processing, Shipped, Completed, Cancelled]
    AdminSummary:
      type: object
      required: [productCount, orderCount, revenue, statusCounts]
      properties:
        productCount:
          type: integer
        orderCount:
          type: integer
        revenue:
          type: number
          format: double
        statusCounts:
          type: object
          additionalProperties:
            type: integer
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string