# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSummaryStatusCounts } from './adminSummaryStatusCounts';

export interface AdminSummary {
  productCount: number;
  orderCount: number;
  revenue: number;
  statusCounts: AdminSummaryStatusCounts;
}