# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  address: string;
  total: number;
  status: OrderStatus;
  createdAt: Date;
  items: OrderItem[];
}