# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  address: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}