# lib/api-zod/src/generated/types/orderItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemInput {
  /** @minimum 1 */
  productId: number;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}