# lib/api-zod/src/generated/types/orderStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusInputStatus } from './orderStatusInputStatus';

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}