# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MarketNest marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  icon?: string;
  imageUrl?: string;
}